requirements.txt
atomicwrites==1.4.0
attrs==20.2.0
colorama==0.4.4
Django==2.2.16
django-cors-headers==3.5.0
django-enumfield==2.0.2
djangorestframework==3.12.1
djangorestframework-api-key==2.0.0
importlib-metadata==2.0.0
iniconfig==1.1.1
packaging==20.4
pip==20.1.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytz==2020.1
setuptools==47.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
zipp==3.4.0
